Implemented PlayerAbilityMessageHandler! Bunch of smaller stuff tweaked...
Signed-off by Zloteanu Nikita<teapa011@gmail.com>

+++ src/main/java/org/spout/vanilla/world/generator/normal/object/tree/BigTreeObject.java
@@ -165,7 +165,7 @@ public class BigTreeObject extends TreeObject {
 		if (layer < totalHeight / 3) {
 			return -1f;
 		} else if (layer == totalHeight / 2) {
-			return totalHeight / 4;
+			return ( (float) totalHeight) / 4;
 		} else if (layer >= totalHeight || layer <= 0) {
 			return 0;
 		} else {
